Fix Back Admin Role

+++ hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/manager/admin/contest/AdminContestManager.java
@@ -90,9 +90,7 @@ public class AdminContestManager {
         // 获取当前登录的用户
         AccountProfile userRolesVo = (AccountProfile) SecurityUtils.getSubject().getPrincipal();
 
-        // 是否为超级管理员或者题目管理或者普通管理
         boolean isRoot = SecurityUtils.getSubject().hasRole("root")
-                || SecurityUtils.getSubject().hasRole("problem_admin")
                 || SecurityUtils.getSubject().hasRole("admin");
 
         // 只有超级管理员和题目管理和训练拥有者才能操作
@@ -158,7 +156,22 @@ public class AdminContestManager {
         return adminContestVo;
     }
 
-    public void deleteContest(Long cid) throws StatusFailException {
+    public void deleteContest(Long cid) throws StatusFailException, StatusForbiddenException {
+        Contest contest = contestEntityService.getById(cid);
+        if (contest == null) { // 查询不存在
+            throw new StatusFailException("查询失败：该比赛不存在,请检查参数cid是否准确！");
+        }
+        // 获取当前登录的用户
+        AccountProfile userRolesVo = (AccountProfile) SecurityUtils.getSubject().getPrincipal();
+
+        boolean isRoot = SecurityUtils.getSubject().hasRole("root")
+                || SecurityUtils.getSubject().hasRole("admin");
+
+        // 只有超级管理员和题目管理和训练拥有者才能操作
+        if (!isRoot && !userRolesVo.getUid().equals(contest.getUid())) {
+            throw new StatusForbiddenException("对不起，你无权限操作！");
+        }
+
         boolean isOk = contestEntityService.removeById(cid);
         /*
          * contest的id为其他表的外键的表中的对应数据都会被一起删除！
@@ -166,7 +179,6 @@ public class AdminContestManager {
         if (!isOk) { // 删除成功
             throw new StatusFailException("删除失败");
         }
-        AccountProfile userRolesVo = (AccountProfile) SecurityUtils.getSubject().getPrincipal();
         log.info("[{}],[{}],cid:[{}],operatorUid:[{}],operatorUsername:[{}]",
                 "Admin_Contest", "Delete", cid, userRolesVo.getUid(), userRolesVo.getUsername());
     }
@@ -244,9 +256,7 @@ public class AdminContestManager {
         // 获取当前登录的用户
         AccountProfile userRolesVo = (AccountProfile) SecurityUtils.getSubject().getPrincipal();
 
-        // 是否为超级管理员或者题目管理或者普通管理
         boolean isRoot = SecurityUtils.getSubject().hasRole("root")
-                || SecurityUtils.getSubject().hasRole("problem_admin")
                 || SecurityUtils.getSubject().hasRole("admin");
 
         // 只有超级管理员和题目管理和训练拥有者才能操作
@@ -311,9 +321,7 @@ public class AdminContestManager {
         // 获取当前登录的用户
         AccountProfile userRolesVo = (AccountProfile) SecurityUtils.getSubject().getPrincipal();
 
-        // 是否为超级管理员或者题目管理或者普通管理
         boolean isRoot = SecurityUtils.getSubject().hasRole("root")
-                || SecurityUtils.getSubject().hasRole("problem_admin")
                 || SecurityUtils.getSubject().hasRole("admin");
 
         // 只有超级管理员和题目管理和训练拥有者才能操作
